# contracts/openapi.yaml
openapi: 3.1.1
info:
  title: Tawsel public contract — identity, intake, planning and execution
  version: 0.1.0
  description: P07–P21 locally implemented identity, intake, locations, planning,
    round start, current activity, outcomes, eligibility, workday closure,
    device takeover and source-branch return receipt/disposition. Provider,
    browser and PostgreSQL evidence are distinguished in phase ledgers. Branch
    resume/redispatch, corrections, full offline capture and signed delivery
    remain designed.
  license:
    name: Private development contract; no public release
...
paths:
...
  /api/v1/intake/task:
    get:
      operationId: intake.getTask
      summary: Read current source snapshot, holder, readiness and dispatch identifiers.
      tags:
        - ERP intake
      security:
        - ProvisioningService: []
      parameters:
        - in: query
          name: externalId
          required: true
          schema:
            type: string
            minLength: 1
            maxLength: 256
      responses:
        "200":
          description: Canonical response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/B2bTask"
        "400": *a9
        "401": *a9
        "403": *a9
        "404": *a9
        "409": *a10
        "422": *a9
        "503": *a9
  /api/v1/intake/tasks:
    get:
      operationId: intake.listTasks
      summary: List source-scoped held/prepared and unassigned/withdrawn work,
        filtered before pagination.
      tags:
        - ERP intake
      security:
        - ProvisioningService: []
      parameters:
        - in: query
          name: state
          required: false
          schema:
            enum:
              - unassigned
              - prepared
              - held
              - withdrawn
        - in: query
          name: driverExternalId
          required: false
          schema:
            type: string
            minLength: 1
            maxLength: 256
        - in: query
          name: limit
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - in: query
          name: cursor
          required: false
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Canonical response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/B2bTaskList"
        "400": *a9
        "401": *a9
        "403": *a9
        "404": *a9
        "409": *a10
        "422": *a9
        "503": *a9
  /api/v1/intake/results/{actionId}:
    get:
      operationId: intake.getBatchResult
      summary: Recover durable accepted/rejected source results; 202 pending means no
        committed result is visible, not proof of receipt.
      tags:
        - ERP intake
      security:
        - ProvisioningService: []
      parameters:
        - in: path
          name: actionId
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Canonical response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/B2bBatchResult"
        "202":
          description: No committed result visible; keep source command pending and retry
            the exact envelope.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/B2bBatchResult"
        "400": *a9
        "401": *a9
        "403": *a9
        "404": *a9
        "409": *a10
        "422": *a9
        "503": *a9
...
  /api/v1/intake/cycles:
    get:
      operationId: dispatch.listCycles
      summary: List preserved source shipment cycles and their frozen snapshots.
      tags:
        - ERP intake
      security:
        - ProvisioningService: []
      parameters:
        - in: query
          name: externalId
          required: true
          schema:
            type: string
            minLength: 1
            maxLength: 256
        - name: cursor
          in: query
          required: false
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Canonical response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/B2bCycleList"
        "400": &a20
          description: Source validation, authorization or dependency error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
            application/json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/ActionResult"
                  - $ref: "#/components/schemas/Problem"
        "401": *a20
        "403": *a20
        "404": *a20
        "409":
          description: Retained business rejection or idempotency conflict
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/ActionResult"
                  - $ref: "#/components/schemas/Problem"
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "422": *a20
        "503": *a20
components:
  schemas:
...
    Problem:
      $ref: ./common.schema.json#/$defs/Problem
...
    ActionResult:
      $ref: ./action-result.v1.schema.json
...
    B2bTask:
      $ref: ./b2b-intake.schema.json#/$defs/Task
    B2bTaskList:
      $ref: ./b2b-intake.schema.json#/$defs/TaskList
    B2bBatchResult:
      $ref: ./b2b-intake.schema.json#/$defs/BatchResult
...
    B2bCycleList:
      $ref: ./b2b-intake.schema.json#/$defs/CycleList
  securitySchemes:
    companySession:
      type: apiKey
      in: cookie
      name: __Host-tawsel-company
    personalSession:
      type: apiKey
      in: cookie
      name: __Host-tawsel-personal
    ProvisioningService:
      type: http
      scheme: bearer
      description: twp_<credential UUID>.<32 random bytes lowercase hex>. Hashed
        server-side; tenant/integration bound, expiring, rotatable. TLS required
        outside loopback.
    ProvisioningOperator:
      type: http
      scheme: bearer
      description: Out-of-band server operator token; bootstrap and recovery rotation
        only. Never distribute to ERP application users.
...